move ScanMetadata to unified-data-interface
