--- src/common/types/store-data/scan-meta-data.ts
@@ -1,12 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-import { ToolData } from 'common/types/store-data/unified-data-interface';
-
-import { TargetAppData } from 'common/types/store-data/unified-data-interface';
-
-export type ScanMetadata = {
-    timestamp: string;
-    toolData: ToolData;
-    targetAppInfo: TargetAppData;
-    deviceName?: string;
-};
