+++ src/tests/unit/tests/electron/report/unified-details-section.test.tsx
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { ScanMetadata } from 'common/types/store-data/scan-meta-data';
-import { TargetAppData } from 'common/types/store-data/unified-data-interface';
+import { ScanMetadata, TargetAppData } from 'common/types/store-data/unified-data-interface';
 import { createDeviceNameItemInfo } from 'electron/views/report/unified-details-section';
 
 describe('UnifiedDetailsSection', () => {
