+++ src/tests/unit/tests/reports/components/report-sections/automated-checks-header-section.test.tsx
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { ScanMetadata } from 'common/types/store-data/scan-meta-data';
+import { ScanMetadata } from 'common/types/store-data/unified-data-interface';
 import { shallow } from 'enzyme';
 import * as React from 'react';
 import { AutomatedChecksHeaderSection } from 'reports/components/report-sections/automated-checks-header-section';
