+++ src/tests/unit/tests/reports/components/report-sections/details-section.test.tsx
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { ScanMetadata } from 'common/types/store-data/scan-meta-data';
+import { ScanMetadata } from 'common/types/store-data/unified-data-interface';
 import * as React from 'react';
 import { getUrlItemInfo } from 'reports/components/report-sections/details-section';
