+++ src/tests/unit/tests/reports/components/report-sections/footer-text.test.tsx
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { ScanMetadata } from 'common/types/store-data/scan-meta-data';
-import { ToolData } from 'common/types/store-data/unified-data-interface';
+import { ScanMetadata, ToolData } from 'common/types/store-data/unified-data-interface';
 import { shallow } from 'enzyme';
 import * as React from 'react';
 import { FooterText } from 'reports/components/report-sections/footer-text';
